Multi Player Game Who wants to be a Millionaire
Developed a secure software programming Game “Who Wants to be a
Millionaire” using JSP, servlets and HTML, CSS. It was a multiplayer
game hosted on localhost in which more than one person can participate
and play. For each question, player can set a betting value and he gets
money depending on the answer. Player with the highest money in the end
wins the game. Technology Used: JSP, Servlets, HTML, CSS, Oracle10g and
Apache-Tomcat 8